Transfer Pickle operations from LoanMailer to a dedicated class
This new class is dedicated to saving loans into an ordinary Python list, which is dumped to disk and loaded using Pickle. This dedicated class prepares the use of concurrent classes for managing loans data.

In the short term, the first concurrent data manager may use Plyvel and LevelDB. The Pickle one is kept for cases where the external dependency of using LevelDB is not wanted.